requirements.txt
gitdb==4.0.7
GitPython==3.1.14
Jinja2==2.11.3
MarkupSafe==1.1.1
smmap==4.0.0
